chore: added TODO comment renamed main file
